Fix bug in applying line style to edge
Newly added edges has 'source' and 'target' fields that were integer
ids, rather than node objects.

These are  automatically converted to objects, but this happens after
restart() is called. Thus, the test for where an edge corresponds to
a mandatory reaction does not work correctly, and mis-identifies a
newly added edge as mandatory, so that  newly drawn edges are all solid
lines, which reset to dashed lines (if necessary) the next time that
restart() is called.